chore: remove not found page component
