+++ apps/web/vibes/soul/examples/pages/not-found/electric.tsx
@@ -1,7 +1,58 @@
 import { copyright, footerLinks } from '@/vibes/soul/examples/sections/footer/electric'
-import { headerLinks } from '@/vibes/soul/examples/sections/header/electric'
-import { NotFoundPage } from '@/vibes/soul/pages/not-found'
+import { headerLinks, logo } from '@/vibes/soul/examples/sections/header/electric'
+import { AnnouncementBar } from '@/vibes/soul/primitives/announcement-bar'
+import { ProductsCarousel } from '@/vibes/soul/primitives/products-carousel'
+import { Footer } from '@/vibes/soul/sections/footer'
+import {
+  Amex,
+  ApplePay,
+  Bitcoin,
+  GooglePay,
+  Mastercard,
+  Paypal,
+  Visa,
+} from '@/vibes/soul/sections/footer/payment-icons'
+import { Facebook, Instagram, X, Youtube } from '@/vibes/soul/sections/footer/social-icons'
+import { Header } from '@/vibes/soul/sections/header'
+import { NotFound } from '@/vibes/soul/sections/not-found'
+import { Subscribe } from '@/vibes/soul/sections/subscribe'
 import { ProductCardProduct } from '@/vibes/soul/types'
+
+const socialMediaLinks = [
+  {
+    href: '#',
+    icon: <Facebook />,
+  },
+  {
+    href: '#',
+    icon: <X />,
+  },
+  {
+    href: '#',
+    icon: <Instagram />,
+  },
+  {
+    href: '#',
+    icon: <Youtube />,
+  },
+]
+
+const paymentIconsArray: React.ReactNode[] = [
+  <Visa key="Visa" />,
+  <Amex key="Amex" />,
+  <Mastercard key="Mastercard" />,
+  <Paypal key="Paypal" />,
+  <GooglePay key="GooglePay" />,
+  <ApplePay key="ApplePay" />,
+  <Bitcoin key="Bitcoin" />,
+]
+
+const locales = [
+  { id: '1', region: 'US', language: 'EN' },
+  { id: '2', region: 'FR', language: 'FR' },
+  { id: '3', region: 'DE', language: 'DC' },
+  { id: '4', region: 'IT', language: 'IT' },
+]
 
 // Products
 export const products: ProductCardProduct[] = [
@@ -117,12 +168,37 @@ export const products: ProductCardProduct[] = [
 
 export default function Preview() {
   return (
-    <NotFoundPage
-      products={products}
-      headerLinks={headerLinks}
-      logo="SOUL"
-      footerLinks={footerLinks}
-      copyright={copyright}
-    />
+    <>
+      <AnnouncementBar>
+        Get <strong>15% off</strong> and free shipping with discount code{' '}
+        <strong>&quot;welcome&quot;</strong>
+      </AnnouncementBar>
+
+      <Header
+        links={headerLinks}
+        logo={logo}
+        cartHref="#"
+        accountHref="#"
+        locales={locales}
+        activeLocale="EN"
+      />
+
+      <NotFound />
+
+      <ProductsCarousel className="mb-6 @xl:mb-8" products={products} />
+
+      <Subscribe
+        title="Sign up for our newsletter"
+        description="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor."
+      />
+
+      <Footer
+        logo={logo}
+        sections={footerLinks}
+        copyright={copyright}
+        paymentIcons={paymentIconsArray}
+        socialMediaLinks={socialMediaLinks}
+      />
+    </>
   )
 }
